Oct 29, 2023, 8:47 PM

+++ xEXERCISE_QUIZ/q1/q1.java
@@ -22,7 +22,7 @@ public class q1 {
 
         while (read.hasNext()) {
             String name = read.next();
-            double score = read.nextInt();
+            int score = read.nextInt();
 
             System.out.println("Name: " + name + "  Score: " + score);
         }
